Update committers.md INFRA-25450 improved display
In the section on 'how do I find out who is subscribed to a mailing list?' the display was poor because of the use of <var> tags, and there was inconsistency in the variables used in the examples. I have cleaned up the display and made the examples more consistent, with an introductory explanation.